requirements.txt
# Flask Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Task Scheduling
APScheduler==3.10.4

# HTTP Requests (für Postal API)
requests==2.31.0

# IMAP Email Checking
imapclient==3.0.1

# OpenAI Integration
openai==1.6.1

# Security & Encryption
cryptography==41.0.7
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.23

# Utilities
python-dateutil==2.8.2

# Production Server
gunicorn==21.2.0
